// src/interfaces/workSpaceInterface.ts
export class workSpaceInterface {
    docs: any[];
    workspaceCategory: string;
    userId: string;
    start: string;
    end: string;
    serviceId: string
    duration: string;
    prepaid: string;
    pass: string;
    passType: string;
    limit: number;
    approved: string;
    bill: string;
    indexOf: any;
    splice: any;

}